+++ src/emqtt_sock.erl
@@ -50,6 +50,7 @@ connect(Host, Port, SockOpts, Timeout) ->
         {ok, Sock} ->
             case lists:keyfind(ssl_opts, 1, SockOpts) of
                 {ssl_opts, SslOpts} ->
+                    ?IS_QoE andalso put(tcp_connected_at, erlang:monotonic_time(millisecond)),
                     ssl_upgrade(Host, Sock, SslOpts, Timeout);
                 false -> {ok, Sock}
             end;
@@ -75,7 +76,23 @@ ssl_upgrade(Host, Sock, SslOpts0, Timeout) ->
 send(Sock, Data) when is_port(Sock) ->
     send_tcp_data(Sock, Data);
 send(#ssl_socket{ssl = SslSock}, Data) ->
-    ssl:send(SslSock, Data);
+    case ssl:send(SslSock, Data) of
+        ok ->
+            ok;
+        {error, closed}->
+            %% We attempt to grab an async exception with more information, if
+            %% available; otherwise, bail out.
+            receive
+                {ssl_error, _Sock, DetailedReason} ->
+                    {error, DetailedReason};
+                {ssl_closed, _Sock} ->
+                    {error, closed}
+            after 1 ->
+                    {error, closed}
+            end;
+        {error, Reason}->
+            {error, Reason}
+    end;
 send(QuicStream, Data) when is_reference(QuicStream) ->
     case quicer:send(QuicStream, Data) of
         {ok, _Len} ->