+++ changelog.md
@@ -1,3 +1,16 @@
+# 1.14.2
+
+- WebSocket: Correctly adopt Gun 2.x interfaces and semantics.
+
+# 1.14.1
+
+- QUIC: Correctly handle peer stream recv abort.
+- TLS1.3: Fix crash after recv session ticket
+
+# 1.14.0
+
+- Switch to using `gun` 2.1.0 as WebSocket client.
+
 # 1.13.7
 
 - Fix a typo in MQTT v3 reason code.